+++ meeting/meetingController.js
@@ -1,17 +1,6 @@
 var User = require('../user/userSchema');
 var Meeting = require('./meetingSchema');
-var nodemailer = require('nodemailer');
-
-var transporter = nodemailer.createTransport({
-    host: 'smtp.yandex.com',
-    port: 465,
-    secure: true,
-    auth: {
-        user: 'TimeetService@yandex.com',
-        pass: 'meetadmin'
-    }
-});
-
+var emailService = require('../services/emailService.js');
 
 
 // Create new meeting
@@ -27,25 +16,7 @@ exports.postMeeting = function (req, res) {
             res.status(400).send(err);
             return;
         }
-        var link = 'http://localhost:8000/#!/scheduling/' + meeting._id;
-
-        var mailOptions = {
-            from: 'TimeetService@yandex.com',
-            to: meeting.participantEmails.toString(),
-            subject: meeting.name,
-            html: '<p>' + meeting.purpose + '</p>'
-            + '<p>Click <a href="'+ link + '">here</a> to set your availabilities</p>'
-        };
-
-        transporter.sendMail(mailOptions, function(error, info){
-            if (error) {
-                console.log(error);
-            } else {
-                console.log('Email sent: ' + info.response);
-            }
-        });
-
-
+        emailService.sendEmails(meeting);
         res.status(201).json(m);
     });
 };
@@ -69,7 +40,6 @@ exports.getMeeting = function (req, res) {
             res.status(400).send(err);
             return;
         }
-
         res.json(meeting);
     });
 };
