+++ services/schedulingService.js
@@ -1,0 +1,7 @@
+module.exports.findBestSlots = function(availabilities) {
+    for (availability in availabilities) {
+        for (slot in availability.slots){
+
+        }
+    }
+}
